feat: added schema builder function
