+++ apps/ledger/src/schema/schema.service.ts
@@ -15,7 +15,6 @@ import { Prisma, schema } from '@prisma/client';
 import {
   ISaveSchema,
   ISchema,
-  ISchemaAttributesFormat,
   ISchemaCredDeffSearchInterface,
   ISchemaExist,
   ISchemaSearchCriteria,
@@ -26,7 +25,6 @@ import {
   ICreateSchema,
   ICreateW3CSchema,
   IGenericSchema,
-  IProductSchema,
   IUpdateSchema,
   IUserRequestInterface,
   UpdateSchemaResponse
@@ -57,16 +55,7 @@ import { networkNamespace } from '@credebl/common/common.utils';
 import { checkDidLedgerAndNetwork } from '@credebl/common/cast.helper';
 import { NATSClient } from '@credebl/common/NATSClient';
 import { from } from 'rxjs';
-import UniqueItems from 'libs/validations/uniqueItems';
-import MaxItems from 'libs/validations/maxItems';
-import MinItems from 'libs/validations/minItems';
-import MultipleOf from 'libs/validations/multipleOf';
-import ExclusiveMinimum from 'libs/validations/exclusiveMinimum';
-import Minimum from 'libs/validations/minimum';
-import Pattern from 'libs/validations/pattern';
-import MaxLength from 'libs/validations/maxLength';
-import MinLength from 'libs/validations/minLength';
-import { IW3CAttributeValue } from '@credebl/common/interfaces/interface';
+import { w3cSchemaBuilder } from 'apps/ledger/libs/helpers/w3c.schema.builder';
 
 @Injectable()
 export class SchemaService extends BaseService {
@@ -319,7 +308,7 @@ export class SchemaService extends BaseService {
         url = `${agentEndPoint}${CommonConstants.SHARED_CREATE_POLYGON_W3C_SCHEMA}${tenantId}`;
       }
 
-      const schemaObject = await this.w3cSchemaBuilder(attributes, schemaName, description);
+      const schemaObject = await w3cSchemaBuilder(attributes, schemaName, description);
       if (!schemaObject) {
         throw new BadRequestException(ResponseMessages.schema.error.schemaBuilder, {
           cause: new Error(),
@@ -368,369 +357,6 @@ export class SchemaService extends BaseService {
       this.logger.error(`[createSchema] - outer Error: ${JSON.stringify(error)}`);
       throw error;
     }
-  }
-
-  private async w3cSchemaBuilder(
-    attributes: IW3CAttributeValue[],
-    schemaName: string,
-    description: string
-  ): Promise<object> {
-    // Function to apply validations based on attribute properties
-    const applyValidations = (attribute, propertyObj): ISchemaAttributesFormat => {
-      const context = { ...propertyObj };
-
-      // Apply string validations
-      if ('string' === attribute.schemaDataType.toLowerCase()) {
-        if (attribute.minLength !== undefined) {
-          const validation = new MinLength(attribute.minLength);
-          validation.json(context);
-        }
-
-        if (attribute.maxLength !== undefined) {
-          const validation = new MaxLength(attribute.maxLength);
-          validation.json(context);
-        }
-
-        if (attribute.pattern !== undefined) {
-          const validation = new Pattern(attribute.pattern);
-          validation.json(context);
-        }
-      }
-
-      // Apply number validations
-      if (['number', 'integer'].includes(attribute.schemaDataType.toLowerCase())) {
-        if (attribute.minimum !== undefined) {
-          const validation = new Minimum(attribute.minimum);
-          validation.json(context);
-        }
-
-        if (attribute.exclusiveMinimum !== undefined) {
-          const validation = new ExclusiveMinimum(attribute.exclusiveMinimum);
-          validation.json(context);
-        }
-
-        if (attribute.multipleOf !== undefined) {
-          const validation = new MultipleOf(attribute.multipleOf);
-          validation.json(context);
-        }
-      }
-
-      // Apply array validations
-      if ('array' === attribute.schemaDataType.toLowerCase()) {
-        if (attribute.minItems !== undefined) {
-          const validation = new MinItems(attribute.minItems);
-          validation.json(context);
-        }
-
-        if (attribute.maxItems !== undefined) {
-          const validation = new MaxItems(attribute.maxItems);
-          validation.json(context);
-        }
-
-        if (attribute.uniqueItems !== undefined) {
-          const validation = new UniqueItems(attribute.uniqueItems);
-          validation.json(context);
-        }
-      }
-
-      return context;
-    };
-
-    // Function to recursively process attributes
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const processAttributes = (attrs: IW3CAttributeValue[]): IProductSchema => {
-      if (!Array.isArray(attrs)) {
-        return { properties: {}, required: [] };
-      }
-
-      const properties = {};
-      const required = [];
-
-      attrs.forEach((attribute, index) => {
-        const { attributeName, schemaDataType, isRequired, displayName } = attribute;
-
-        // Add to required array if isRequired is true
-        if (isRequired) {
-          required.push(attributeName);
-        }
-
-        // Create base property object with common fields
-        const baseProperty = {
-          type: schemaDataType.toLowerCase(),
-          order: index,
-          title: displayName || attributeName,
-          description: `${attributeName} field`
-        };
-
-        // Handle different attribute types
-        if (['string', 'number', 'boolean', 'integer'].includes(schemaDataType.toLowerCase())) {
-          // Apply validations to the base property
-          properties[attributeName] = applyValidations(attribute, baseProperty);
-        } else if ('datetime-local' === schemaDataType.toLowerCase()) {
-          properties[attributeName] = {
-            ...baseProperty,
-            type: 'string',
-            format: 'date-time'
-          };
-        } else if ('array' === schemaDataType.toLowerCase() && attribute.items) {
-          const result = processAttributes(attribute.items);
-
-          properties[attributeName] = {
-            ...baseProperty,
-            type: 'array',
-            items: {
-              type: 'object',
-              properties: result.properties
-            }
-          };
-
-          // Apply array-specific validations
-          properties[attributeName] = applyValidations(attribute, properties[attributeName]);
-
-          // Add required properties to the items schema if any
-          if (0 < result.required.length) {
-            properties[attributeName].items.required = result.required;
-          }
-        } else if ('object' === schemaDataType.toLowerCase() && attribute.properties) {
-          const nestedProperties = {};
-          const nestedRequired = [];
-
-          // Process each property in the object
-          Object.keys(attribute.properties).forEach((propKey) => {
-            const prop = attribute.properties[propKey];
-
-            // Add to nested required array if isRequired is true
-            if (prop.isRequired) {
-              nestedRequired.push(propKey);
-            }
-
-            // Create base property for nested object
-            const nestedBaseProperty = {
-              type: prop.schemaDataType.toLowerCase(),
-              title: prop.displayName || prop.attributeName,
-              description: `${prop.attributeName} field`
-            };
-
-            if ('array' === prop.schemaDataType.toLowerCase() && prop.items) {
-              // Handle nested arrays
-              const result = processAttributes(prop.items);
-
-              nestedProperties[propKey] = {
-                ...nestedBaseProperty,
-                type: 'array',
-                items: {
-                  type: 'object',
-                  properties: result.properties
-                }
-              };
-
-              // Apply array-specific validations
-              nestedProperties[propKey] = applyValidations(prop, nestedProperties[propKey]);
-
-              // Add required properties to the items schema if any
-              if (0 < result.required.length) {
-                nestedProperties[propKey].items.required = result.required;
-              }
-            } else {
-              // Handle basic properties with validations
-              nestedProperties[propKey] = applyValidations(prop, nestedBaseProperty);
-            }
-          });
-
-          properties[attributeName] = {
-            ...baseProperty,
-            type: 'object',
-            properties: nestedProperties
-          };
-
-          // Add required properties to the object schema if any
-          if (0 < nestedRequired.length) {
-            properties[attributeName].required = nestedRequired;
-          }
-        }
-      });
-
-      return { properties, required };
-    };
-
-    // Process all attributes
-    const result = processAttributes(attributes);
-    const { properties } = result;
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const required = ['id', ...result.required];
-
-    // Add id property
-    properties['id'] = {
-      type: 'string',
-      format: 'uri'
-    };
-
-    const date = new Date().toISOString();
-    const schemaNameObject = {};
-    schemaNameObject[schemaName] = {
-      const: schemaName
-    };
-
-    const W3CSchema = {
-      $schema: 'https://json-schema.org/draft/2020-12/schema',
-      $id: `${date}-${schemaName}`,
-      $vocabulary: {
-        'https://json-schema.org/draft/2020-12/vocab/core': true,
-        'https://json-schema.org/draft/2020-12/vocab/applicator': true,
-        'https://json-schema.org/draft/2020-12/vocab/unevaluated': true,
-        'https://json-schema.org/draft/2020-12/vocab/validation': true,
-        'https://json-schema.org/draft/2020-12/vocab/meta-data': true,
-        'https://json-schema.org/draft/2020-12/vocab/format-annotation': true,
-        'https://json-schema.org/draft/2020-12/vocab/content': true
-      },
-      type: 'object',
-      required: ['@context', 'issuer', 'issuanceDate', 'type', 'credentialSubject'],
-      properties: {
-        '@context': {
-          $ref: '#/$defs/context'
-        },
-        type: {
-          type: 'array',
-          items: {
-            anyOf: [
-              {
-                $ref: '#/$defs/VerifiableCredential'
-              },
-              {
-                const: `#/$defs/$${schemaName}`
-              }
-            ]
-          }
-        },
-        credentialSubject: {
-          $ref: '#/$defs/credentialSubject'
-        },
-        id: {
-          type: 'string',
-          format: 'uri'
-        },
-        issuer: {
-          $ref: '#/$defs/uriOrId'
-        },
-        issuanceDate: {
-          type: 'string',
-          format: 'date-time'
-        },
-        expirationDate: {
-          type: 'string',
-          format: 'date-time'
-        },
-        credentialStatus: {
-          $ref: '#/$defs/credentialStatus'
-        },
-        credentialSchema: {
-          $ref: '#/$defs/credentialSchema'
-        }
-      },
-      $defs: {
-        context: {
-          type: 'array',
-          items: [
-            {
-              const: 'https://www.w3.org/2018/credentials/v1'
-            }
-          ],
-          additionalItems: {
-            oneOf: [
-              {
-                type: 'string',
-                format: 'uri'
-              },
-              {
-                type: 'object'
-              },
-              {
-                type: 'array',
-                items: {
-                  $ref: '#/$defs/context'
-                }
-              }
-            ]
-          },
-          minItems: 1,
-          uniqueItems: true
-        },
-        credentialSubject: {
-          type: 'object',
-          required: ['id'],
-          additionalProperties: false,
-          properties
-        },
-        VerifiableCredential: {
-          const: 'VerifiableCredential'
-        },
-        credentialSchema: {
-          oneOf: [
-            {
-              $ref: '#/definitions/idAndType'
-            },
-            {
-              type: 'array',
-              items: {
-                $ref: '#/definitions/idAndType'
-              },
-              minItems: 1,
-              uniqueItems: true
-            }
-          ]
-        },
-        credentialStatus: {
-          oneOf: [
-            {
-              $ref: '#/definitions/idAndType'
-            },
-            {
-              type: 'array',
-              items: {
-                $ref: '#/definitions/idAndType'
-              },
-              minItems: 1,
-              uniqueItems: true
-            }
-          ]
-        },
-        idAndType: {
-          type: 'object',
-          required: ['id', 'type'],
-          properties: {
-            id: {
-              type: 'string',
-              format: 'uri'
-            },
-            type: {
-              type: 'string'
-            }
-          }
-        },
-        uriOrId: {
-          oneOf: [
-            {
-              type: 'string',
-              format: 'uri'
-            },
-            {
-              type: 'object',
-              required: ['id'],
-              properties: {
-                id: {
-                  type: 'string',
-                  format: 'uri'
-                }
-              }
-            }
-          ]
-        },
-        ...schemaNameObject
-      },
-      title: schemaName,
-      description: `${description}`
-    };
-    return W3CSchema;
   }
 
   private async storeW3CSchemas(schemaDetails, user, orgId, attributes, alias): Promise<schema> {
